V0.106.2 more changes to the  way the site is structured.
