+++ src/assets/sources.json
@@ -91,7 +91,7 @@
     }
   ],
 
-  "/chapter4.1": [
+  "/chapter4.2": [
     {
       "author": "A Nikkei Company",
       "year": "2023, september 12",
