+++ src/assets/terms.json
@@ -35,17 +35,17 @@
   {
     "term": "Semigestructureerd interview",
     "definition": "Een interviewvorm waarbij de interviewer een lijst met vragen heeft, maar ook ruimte laat om spontan vragen te stellen. Op die manier kan je dieper ingaan op de antwoorden van de persoon die je interviewt. Hiedoor krijg je meer informatie dan wanneer je bijvoorbeeld een enquête afneemt.",
-    "page": "2.1 Onderzoeken"
+    "page": "4.1 Inleven en eerste interviews audio gebruikers"
   },
   {
     "term": "SBO",
     "definition": "Een school voor speciaal basisonderwijs. Op deze scholen kan men goed omgaan met kinderen met gedragsproblemen of leerproblemen. Dit zijn kinderen die niet goed kunnen meekomen in het reguliere onderwijs, maar die ook geen speciaal onderwijs nodig hebben.",
-    "page": "2.1 Onderzoeken"
+    "page": "4.1 Inleven en eerste interviews audio gebruikers"
   },
   {
     "term": "SO",
     "definition": "Speciaal Onderwijs. Dit is een type onderwijs dat speciaal is ingericht voor kinderen met een beperking. Dit kan een lichamelijke beperking zijn, een verstandelijke beperking, een zintuiglijke beperking of een combinatie hiervan. Ook leerlingen met ernstige gedragsproblemen of langdurige ziekte kunnen terecht op het speciaal onderwijs.",
-    "page": "2.1 Onderzoeken"
+    "page": "4.1 Inleven en eerste interviews audio gebruikers"
   },
   {
     "term": "Digitaal Toegankelijk",
@@ -55,12 +55,12 @@
   {
     "term": "Leber",
     "definition": "Leber LOA, LHON of de ziekte van Leber is een erfelijke aandoening die door schade aan oogzenuwen zichtproblemen veroorzaakt. Dit kan variëren van verminderd zicht tot complete blindheid. Ongeveer 2,3 op de 100.000 mensen in Nederland heeft deze aandoening. Voor onduidelijke redenen kan alleen de moeder het gen doorgeven (ook als de vader het gen heeft). Dit betekent dat alleen kinderen van moeders met deze aandoening het ook kunnen krijgen.",
-    "page": "2.1 Onderzoeken"
+    "page": "4.1 Inleven en eerste interviews audio gebruikers"
   },
   {
     "term": "Amaurosis congenita van Leber",
     "definition": "Amaurosis congenita van Leber of LCA is een erfelijke aandoening die vaak bij de geboorte al zichtbaar is. Het kan leiden tot een ernstige visuele beperking of zelfs blindheid. Het netvlies werkt niet naar behoren waardoor het zicht ernstig kan worden aangetast. In de meeste gevallen krijgt men deze aandoening wanneer beide ouders het gen doorgeven. Helaas is LCA nog niet te genezen.",
-    "page": "2.1 Onderzoeken"
+    "page": "4.1 Inleven en eerste interviews audio gebruikers"
   },
   {
     "term": "WCAG",
@@ -69,6 +69,6 @@
   {
     "term": "Alt Tekst",
     "definition": "Een vervangende tekst die wordt voorgelezen door een schermlezer. Op die manier kan informatie die in een afbeelding staat toch worden overgebracht aan bijvoorbeeld braillegebruikers. Alt tekst wordt ook weergegeven wanneer een afbeelding niet kan worden geladen. En zoekmachines zoals Google gebruiken het ook om de inhoud van een afbeelding te begrijpen.",
-    "page": "2.2 Ontwerpen"
+    "page": "4.2 Iteratie 1 ontwerp schetsen"
   }
 ]
